feature: keep SQS receipt handle in EsNotification
When processing notifications with SQS-triggered AWS Lambda, successful Lambda execution deletes the processed notification from the queue.

It is not the case, though, when you receive a notification from SQS in non-Lambda runtime.
Then you should explicitly acknowledge the notification by deleting it from the queue.

This is where the receipt handle is used.